feat(StreetFighter) : Separar a solution em camadas.

+++ src/modulo-05-.NET-I/ExercicioStreetFighter/StreetFighter.Web/Models/FichaTecnicaModel.cs
@@ -12,12 +12,9 @@ namespace StreetFighter.Web.Models
         public String Nome { get; set; }
         public DateTime DataNascimento { get; set; }
         public int Altura { get; set; }
-        public int Peso { get; set; }
-
+        public decimal Peso { get; set; }
         public String Origem { get; set; }
-
         public String GolpesEspeciais { get; set; }
-
         [DisplayName("Personagem Oculto")]
         public bool PersonagemOculto { get; set; }
